Add Javadoc for F#main
Javadoc is a standard for documenting Java code.
It is based on comments wrapped within '/**' and
'*/'. The full Javadoc reference can be found on
the Internet.
In theory, you should write Javadoc for every
class and method. In practice, write code with
clear class, method and variable names to reduce
the need for spending time documenting
everything.
I added this one because the linter asked for it;
it probably didn't ask for it for
F#fibonaccirecursivo because it is private.