Make line too long shorter
This kind of code style rule was originally used
because monitors were too narrow so, if you made
lines too long, people would have to scroll
horizontally to read them, which is annoying.
Nowadays monitors fit way more than enough pixels
to draw at least 150 characters in a single line,
but IDEs are so customizable that it is common
to have one developer complaining about whatever
the limit is set to being too low because their
IDE layout puts a bunch of windows next to each
other, lowering the space available for text.
Anyway, the linter we're using sets this to 80, so
this commit fixes that for the problematic line.