Create assertion workflow
Workflows are GitHub's "Continuous Integration and
Delivery" concept. They're pieces of code that
can be run tied to certain events, typically Git
events, such as a push, as is the case with the
ones configured here.
Continuous Integration is the practice of running
some checks on every change to ensure that nothing
breaks. To simplify things, and because it is how
most companies do it even though it isn't really
"continuous", the workflow defined in this commit
will run every time a branch is pushed, on the
latst commmit of that branch.
For now, the workflow has 4 steps: one for
linting (ensuring the code looks according to a
certain set of rules), assembling, generating
javadoc and running tests, but the last three
do nothing yet - you'll write them yourself later
on. The first job of the workflow uses default
rules set by GitHub which I haven't bothered
changing, so we'll have to stick to them -
the previous commits are there because they were
required to fix violations of said rules that
were already present in the codebase.
Continuous Delivery is a similar thing, but
instead of running things to ensure that nothing
breaks, you deliver: package your program and
send it to the customer, an app store, or the
like, optionally with documentation, etc.
Unlike CI, there is not too much value on
running CI on every push as opposed to, for
example, shared references, such as master in our
case. Later on you'll setup CD to package your
code and release it on GitHub where a new commit
lands on master.