Add package-info.java for org.jose
package-info.java is a special file that can be
included in a package to apply things to all
classes inside of it. For example, to annotate
all classes in a package, you can instead create
a package-info.java in that package and annotate
the package statement:

@MyAnnotation
package org.jose;

I haven't seen package-info.java being used in
pretty much any of the projects I've worked on.
Very likely that the first developers you'll
meet won't even know what it is, and I don't
think you need to care much about either, other
than because the linter we use in this project
forces it.